+++ ft_printf.h
@@ -9,7 +9,7 @@ int	ft_printf(const char *, ...);
 int	ft_isvalidparam(char c);
 int	ft_print_char(int c);
 int	ft_print_hex(int placeholder, int i);
-int	ft_print_ptr(unsigned int i);
+int	ft_print_ptr(unsigned long int i);
 int	ft_print_int(int i);
 int	ft_print_unint(unsigned int u);
 int	ft_print_str(char *s);
